refactor(commentaires): déplacer les composants génériques vers components/commentaires/
Tous les composants devenus agnostiques (post-migration vers le Context)
sortent du namespace `indicateurs/` : BadgeStatut, CarteCommentaire,
ContenuRepliable, EditeurCommentaire, libelleAuteur, LigneHistorique,
ListeCommentaires, meteo, MeteoTag, SelecteurMeteo, SectionSyntheseDesResultats,
SectionAutresCommentaires.

Restent dans `components/indicateurs/commentaires/` les deux fichiers
spécifiques au sujet indicateur : `IndicateurCommentaireConfigProvider`
et `IndicateurCommentairesTab`.

Refs PIL-1588.
