+++ apps/mb-webapp/src/components/commentaires/EditeurCommentaire.tsx
@@ -8,9 +8,9 @@ import { z } from 'zod'
 
 import { useCommentaireConfig } from '@/components/commentaires/CommentaireConfigContext'
 import { EditeurRiche } from '@/components/editeur-riche/EditeurRiche'
-import { BadgeStatut } from '@/components/indicateurs/commentaires/BadgeStatut'
-import { libelleAuteur } from '@/components/indicateurs/commentaires/libelleAuteur'
-import { SelecteurMeteo } from '@/components/indicateurs/commentaires/SelecteurMeteo'
+import { BadgeStatut } from '@/components/commentaires/BadgeStatut'
+import { libelleAuteur } from '@/components/commentaires/libelleAuteur'
+import { SelecteurMeteo } from '@/components/commentaires/SelecteurMeteo'
 import { Button } from '@/components/ui/Button'
 import { Text } from '@/components/ui/Typography'
 import { clsxm } from '@/lib/clsxm'
