+++ apps/mb-webapp/src/components/commentaires/LigneHistorique.tsx
@@ -2,9 +2,9 @@ import { type CommentaireApiModel } from '@pilote/mb-shared/commentaire'
 import { useSuspenseQuery } from '@tanstack/react-query'
 
 import { useCommentaireConfig } from '@/components/commentaires/CommentaireConfigContext'
-import { ContenuRepliable } from '@/components/indicateurs/commentaires/ContenuRepliable'
-import { libelleAuteur } from '@/components/indicateurs/commentaires/libelleAuteur'
-import { couleurIndice, meteoFromIndice } from '@/components/indicateurs/commentaires/meteo'
+import { ContenuRepliable } from '@/components/commentaires/ContenuRepliable'
+import { libelleAuteur } from '@/components/commentaires/libelleAuteur'
+import { couleurIndice, meteoFromIndice } from '@/components/commentaires/meteo'
 import { Text } from '@/components/ui/Typography'
 import { clsxm } from '@/lib/clsxm'
 import { formatDateHeureFr } from '@/lib/format'
