+++ apps/mb-webapp/src/components/commentaires/ListeCommentaires.tsx
@@ -3,9 +3,9 @@ import { CheckCircle2, Eye, EyeOff, History, Pencil, Plus } from 'lucide-react'
 import { type ReactNode, useState } from 'react'
 
 import { useCommentaireConfig } from '@/components/commentaires/CommentaireConfigContext'
-import { CarteCommentaire } from '@/components/indicateurs/commentaires/CarteCommentaire'
-import { EditeurCommentaire } from '@/components/indicateurs/commentaires/EditeurCommentaire'
-import { LigneHistorique } from '@/components/indicateurs/commentaires/LigneHistorique'
+import { CarteCommentaire } from '@/components/commentaires/CarteCommentaire'
+import { EditeurCommentaire } from '@/components/commentaires/EditeurCommentaire'
+import { LigneHistorique } from '@/components/commentaires/LigneHistorique'
 import { Button } from '@/components/ui/Button'
 import { EmptyState } from '@/components/ui/EmptyState'
 import { Text } from '@/components/ui/Typography'
